feat: put video display to a new thread
